pyproject.toml
[project]
name = "quiz-solver-agent"
version = "1.0.0"
description = "Autonomous quiz-solving agent with LangGraph"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.0.40",
    "langchain-core>=0.1.0",
    "playwright>=1.40.0",
    "pytesseract>=0.3.10",
    "speechrecognition>=3.10.0",
    "pydub>=0.25.1",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pillow>=10.0.0",
    "pandas",
    "numpy",
    "pdfplumber",
    "pyPDF2",
    "langchain-google-genai"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"